trivial: SVG badges in README
Two of the three badges at the top of the README were using the PNG image version instead of the SVG version. This made them look a little strange and blurry in comparison when viewing on a Retina/HiDPI screen.

This switches to the SVG versions, which look cleaner and are likely slightly smaller file size anyhow.

Totally trivial, but it looks nicer. :-)