tell travis-ci to use gomodules on >=1.11
Since travis-ci puts files in a constructed GOPATH by default, it will currently not bother with a detected go.mod (hence the `warning: modules disabled by GO111MODULE=auto in GOPATH/src`) message in install phase output.

This is fine and things still build, but it's more robust to test using the defined modules. This change signals to use modules even when within a GOPATH, so that travis-ci will test with them.