tell appveyor to use gomodules when in GOPATH
Same as previous commit but for appveyor support.  In this case it looks like the script is manually constructing the GOPATH. Since the appveyor version is currently only testing on stable 1.11, this change will enable the use of gomodules while in the gopath.

(In the future if desired, since it only tests on stable and greater, the appveyor script could be modified to not have a GOPATH at all, and this env variable would not be necessary. But for now, doing the smallest possible change.)