Add the ability to skip aliased networks
when using Networks and NetworksWithin.

This is probably what most people wanted in the first place. Ideally,
this would be the default behavior, but that is potentially a breaking
change for people.

The code is more complicated than would otherwise be necessary as I am
trying to support the old behavior exactly as it was.