Handle error conditions better in NetworksWithin
Previously, NetworksWithin would allow you to look up an IPv6 network in
an IPv4 database. In this case it would return invalid data.

As the method doesn't have an error return value, it will now return a
Networks with Err() set. When the Networks has an error, iteration will
stop and the Network call will fail.